// client/src/models/Profile.ts
export interface Profile {
  age: string;
  creditScore: string;
  monthlyIncome: string;
  monthlyDebts: string;
  delinquencies30: string;
  delinquencies60: string;
  delinquencies90: string;
  openedLines: string;
  unsecuredUsage: string;
  unsecuredLimit: string;
  hasIncomeVerification: boolean;
  hasBankruptcy: boolean
}